added delete() method to delete hitbox and stuff when leaving world

+++ src/entities/enemy/RedBloon.java
@@ -1,17 +1,16 @@
 package entities.enemy;
 
 import entities.Entity;
-import entities.ExampleActor;
 import entities.tower.Tower;
 
 public class RedBloon extends Enemy {
-    public RedBloon(int speed, int lives) {
+    public RedBloon(double speed, int lives) {
         super(speed, lives);
         setImage("Coin.png"); //TODO FIX @elias/art guys
     }
 
     public void onHit(Entity hitter) {
-        if(hitter instanceof Tower) // TODO REPLACE WITH PROJECTILE @ELIAS
+        if (hitter instanceof Tower) // TODO REPLACE WITH PROJECTILE @ELIAS
         {
             lives--;
             if (lives <= 0) {
